MiddleColumn websocket handlers
Added useWebsocket to MiddleColumn. Added form. Added input onChange
handler (handleTyping) to send a message to the websocket server so the
server can indicate to others who is typing. Added form onSubmit
handler (handleSendMessage) so that strings or files can be uploaded to
the server. This may be changed so only strings can be uploaded in the
future. Added dataTransferObjects types file to manage DTOs sent via
websocket. Also modified TypingIndicator from returning a fragment to
returning a div.