Added MessageList component, types, and fetch
MessageList will retrieve messages via fetch to the api. When the
component mounts and message history is being fetched, websocket will
store received messages in a buffer. When fetching is done, the
messages in the buffer will be concatenated to the fetched messages and
the combined result will be displayed.