Added filter properties to useWebsocket
filter controls whether lastMessage state variable provided by
useWebsocket is updated. When updated, this causes rerenders. By
returning false when messages are received that are irrelevant for
rendering a particular component, we greatly reduce unnecessary
rerenders. We also wrap the filter function in a try/catch block to
catch errors from trying to parse e.data when it is not JSON.