MessageForm no longer spams typing MessageEvents
MessageForm has been updated to track the last typing message sent
using a useRef variable. When the user types, a message is sent to tell
the server that a user is or isn't typing. Now, it compares the typing
variable that would be sent with the one stored in useRef and only
sends a message if they differ. Therefore, a message is only sent to
the websocket server when the length of textarea.value changes from 0
to 1 or 1 to 0, instead of on every keypress.