Corrected type, new MessageList handling
MessageList will now check how much time has elapsed between
consecutive messages from the same user. If 5 minutes have passed, the
message will not be collapsed into the previous one. This is similar to
how Discord handles consecutive messages from the same user.
The IMessageBase type has been updated to reflect the fact the date
property will be a string and not a Date object and will have to be
reconstructed.